backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
openai==1.12.0
supabase==2.3.4
python-dotenv==1.0.1
pandas==2.2.0
python-multipart==0.0.9
pypdf2==3.0.1
openpyxl==3.1.2
